+++ src/larcv3/app/queueio/BatchFillerBBox.h
@@ -53,7 +53,7 @@ namespace larcv3 {
   private:
 
     size_t set_data_size(const EventBBox<dimension>& image_data);
-    int _check_projection(const int & projection_id);
+    int _check_projection(const size_t & projection_id);
 
     std::string _bbox_producer;
     size_t _max_boxes;
@@ -101,4 +101,3 @@ namespace larcv3 {
 
 #endif
 /** @} */ // end of doxygen group
-
