+++ src/larcv3/app/queueio/BatchFillerSparseTensor.h
@@ -53,7 +53,7 @@ namespace larcv3 {
   private:
 
     size_t set_data_size(const EventSparseTensor<dimension>& image_data);
-    int _check_projection(const int & projection_id);
+    int _check_projection(const size_t & projection_id);
 
     std::string _tensor_producer;
     size_t _max_voxels;
@@ -104,4 +104,3 @@ namespace larcv3 {
 
 #endif
 /** @} */ // end of doxygen group
-
